feat: Collector execution
Implemented mock CLI interface as rhc-insights. The 'ls' and 'run'
subcommands are routed and runnable, though '--format=json' doesn't
work yet.

Collector configuration has been altered to match the requirements
and to be able to actually run the collector.

The idea of custom errors has been abandoned (for now) -- we should
log events as soon as they happen, and that should be separate from
the (localized) message returned to the caller method.

Mock collector 'org.example.mock' is now included to simplify testing
and development.